Abstracted the functions into a library to be able to fix #49 in Multi-Go
